// x-pack/platform/packages/shared/response-ops/alerting-v2-episodes-ui/hooks/use_alerting_rules_cache.d.ts
import type { HttpStart } from '@kbn/core-http-browser';
export interface UseAlertingRulesCacheOptions {
    ruleIds: string[];
    services: {
        http: HttpStart;
    };
}
/**
 * Provides a rules cache by id, fetching uncached rules
 * with the minimum number of bulk requests possible.
 * Returns rulesCache as state so consumers re-render when rules are loaded.
 */
export declare const useAlertingRulesCache: ({ ruleIds, services }: UseAlertingRulesCacheOptions) => {
    rulesCache: Record<string, {
        kind: "alert" | "signal";
        metadata: {
            name: string;
            description?: string | undefined;
            owner?: string | undefined;
            tags?: string[] | undefined;
        };
        time_field: string;
        schedule: {
            every: string;
            lookback?: string | undefined;
        };
        evaluation: {
            query: {
                base: string;
            };
        };
        id: string;
        enabled: boolean;
        createdBy: string | null;
        createdAt: string;
        updatedBy: string | null;
        updatedAt: string;
        recovery_policy?: {
            type: "query" | "no_breach";
            query?: {
                base?: string | undefined;
            } | undefined;
        } | undefined;
        state_transition?: {
            pending_operator?: "AND" | "OR" | undefined;
            pending_count?: number | undefined;
            pending_timeframe?: string | undefined;
            recovering_operator?: "AND" | "OR" | undefined;
            recovering_count?: number | undefined;
            recovering_timeframe?: string | undefined;
        } | null | undefined;
        grouping?: {
            fields: string[];
        } | undefined;
        no_data?: {
            behavior?: "recover" | "no_data" | "last_status" | undefined;
            timeframe?: string | undefined;
        } | undefined;
        artifacts?: {
            id: string;
            type: string;
            value: string;
        }[] | undefined;
    }>;
    loading: boolean;
    error: Error | undefined;
};